include: wifi_parport: Fix incorrect type

+++ src/include/wifi_parport.h
@@ -6,6 +6,6 @@
 
 void wifi_parport_prepare(void);
 uint16_t wifi_parport_read_byte(void);
-uint8_t wifi_parport_write_byte(uint8_t val);
+uint16_t wifi_parport_write_byte(uint8_t val);
 
 #endif // __WIFI_PARPORT_H__
